requirements.txt
dataclass-wizard==0.22.2
pika==1.3.1
from-dict==0.3.3
paho-mqtt==1.6.1
fpdf==1.7.2
pytz==2022.7.1
pydantic==1.10.7